requirements.txt
llama-cpp-python==0.2.90
sentence-splitter==1.4